% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banzhaf_appro.R
\name{banzhaf_appro}
\alias{banzhaf_appro}
\title{Banzhaf Index (approximated)}
\usage{
banzhaf_appro(characteristic_func, n_players, n_rep, replace = TRUE)
}
\arguments{
\item{characteristic_func}{The valued function defined on the subsets of the number
of players}

\item{n_players}{The number of players}

\item{n_rep}{The number of iterations to perform in the approximated
calculation}

\item{replace}{should sampling be with replacement?}
}
\value{
The Shapley value for each player
}
\description{
Calculate the approximated Banzhaf Index based on sampling
}
