% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banzhaf_appro.R
\name{banzhaf_appro_func}
\alias{banzhaf_appro_func}
\title{Banzhaf value (approximation)}
\usage{
banzhaf_appro_func(value_func, n_rep, n_players, replace = TRUE)
}
\arguments{
\item{value_func}{The valued function defined on the subsets of the number
of players}

\item{n_rep}{The number of iterations to perform in the approximated
calculation}

\item{n_players}{Only used if \code{value_func} is a \code{function}.
The number of players in the game.}

\item{replace}{should sampling be with replacement?}
}
\value{
The Shapley value for each player
}
\description{
Calculate the approximated Shapley value based on sampling
}
