% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banzhaf_appro.R
\name{banzhaf_appro_vector}
\alias{banzhaf_appro_vector}
\title{Banzhaf Index (approximated)}
\usage{
banzhaf_appro_vector(value_func, n_rep)
}
\arguments{
\item{value_func}{The valued function defined on the subsets of the number
of players}

\item{n_rep}{The number of iterations to perform in the approximated
calculation}
}
\value{
The Banzhaf Index for each player
}
\description{
Calculate the approximated Banzhaf Index based on sampling
}
