% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banzhaf_exact.R
\name{banzhaf_exact}
\alias{banzhaf_exact}
\title{Banzhaf Index (exact)}
\usage{
banzhaf_exact(characteristic_func, n_players)
}
\arguments{
\item{characteristic_func}{The valued function defined on the subsets of the number
of players}

\item{n_players}{The number of players in the game.}
}
\value{
The Banzhaf Index for each player
}
\description{
Calculate the approximated Banzhaf Index
}
