% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banzhaf_exact.R
\name{banzhaf_exact_vector}
\alias{banzhaf_exact_vector}
\title{Banzhaf Index (exact)}
\usage{
banzhaf_exact_vector(value_func)
}
\arguments{
\item{value_func}{The valued function defined on the subsets of the number
of players}
}
\value{
The Banzhaf Index for each player
}
\description{
Calculate the approximated Banzhaf Index
}
