% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banzhaf_owen_exact.R
\name{banzhaf_owen_exact}
\alias{banzhaf_owen_exact}
\title{Banzhaf-Owen Value}
\usage{
banzhaf_owen_exact(characteristic_func, union, n_players)
}
\arguments{
\item{characteristic_func}{The valued function defined on the subsets of the number
of players}

\item{union}{List of vectors indicating the a priori unions between the
players}

\item{n_players}{The number of players in the game.}
}
\value{
The Banzhaf Index for each player
}
\description{
Calculate the approximated Banzhaf-Owen value
}
