% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalitions.R
\name{coalitions}
\alias{coalitions}
\title{coalitions}
\usage{
coalitions(n_players)
}
\arguments{
\item{n_players}{Number of players}
}
\value{
A list containing a \code{data.frame} of the binary representation
of the coalitions and a \code{vector} of the classical representation (as
sets) of the coalitions
}
\description{
Create all the possible coalitions given the number of players
}
