% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_exact.R
\name{core_exact}
\alias{core_exact}
\title{Exact core of the game}
\usage{
core_exact(characteristic_func, n_players = 0)
}
\arguments{
\item{characteristic_func}{The valued function defined on the subsets of the number
of players.}

\item{n_players}{Only used if \code{characteristic_func} is a \code{function}.
The number of players in the game.}
}
\value{
The vertices of the core
}
\description{
Calculate the vertices of core of the game.
}
