% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owen_appro.R
\name{owen_appro}
\alias{owen_appro}
\title{Owen value (approximation)}
\usage{
owen_appro(characteristic_func, union, n_players, n_rep)
}
\arguments{
\item{characteristic_func}{The valued function defined on the subsets of the number
of players}

\item{union}{List of vectors indicating the a priori unions between the
players}

\item{n_players}{The number of players}

\item{n_rep}{The number of iterations to perform in the approximated
calculation}
}
\value{
The Owen value for each player
}
\description{
Calculate the approximated Owen value based on sampling using
the algorithm proposed by Saavedra-Nieves et al. (2018).
}
\references{
Saavedra-Nieves, A., García-Jurado, I., & Fiestras-Janeiro, M. G. (2018).
Estimation of the Owen value based on sampling. In The mathematics of the uncertain:
A tribute to Pedro Gil (pp. 347-356). Cham: Springer International Publishing.
}
