% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owen_exact.R
\name{owen_exact}
\alias{owen_exact}
\title{Owen value (exact)}
\usage{
owen_exact(characteristic_func, union, n_players = NULL)
}
\arguments{
\item{characteristic_func}{The valued function defined on the subsets of the number
of players}

\item{union}{List of vectors indicating the a priori unions between the
players}

\item{n_players}{The number of players}
}
\value{
The Owen value for each player
}
\description{
Calculate the exact Owen
}
