% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predecessor.R
\name{predecessor}
\alias{predecessor}
\title{Predecessor}
\usage{
predecessor(permutation, player, include_player = FALSE)
}
\arguments{
\item{permutation}{A permutation of the players}

\item{player}{Number of the player i}

\item{include_player}{Whether the player i is included as predecessor of
itself or not}
}
\value{
The set of predecessors of the player i in the order 0
}
\description{
Given a permutation 0 of players and a player i, calculate the
set of predecessors of the player i in the order 0
}
