% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_appro.R
\name{shapley_appro_func}
\alias{shapley_appro_func}
\title{Shapley value (approximation)}
\usage{
shapley_appro_func(value_func, n_rep, n_players)
}
\arguments{
\item{value_func}{The valued function defined on the subsets of the number
of players}

\item{n_rep}{The number of iterations to perform in the approximated
calculation}
}
\value{
The Shapley value for each player
}
\description{
Calculate the approximated Shapley value based on sampling
}
