% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_exact.R
\name{shapley_exact_vector}
\alias{shapley_exact_vector}
\title{Shapley value (exact)}
\usage{
shapley_exact_vector(value_func)
}
\arguments{
\item{value_func}{The valued function defined on the subsets of the number
of players}
}
\value{
The Shapley value for each player
}
\description{
Calculate the exact Shapley value
}
