% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banzhaf.R
\name{banzhaf}
\alias{banzhaf}
\title{Banzhaf value}
\usage{
banzhaf(value_func, method = "exact", n_rep = 10000, n_players = 0)
}
\arguments{
\item{value_func}{The valued function defined on the subsets of the number
of players}

\item{method}{Method used to calculate the Banzhaf value. Valid methods are:
\code{exact} for the exact value or \code{appro} for approximated polynomial
calculation based on sampling.}

\item{n_rep}{Only used if \code{method} is \code{appro}. The number of
iterations to perform in the approximated calculation}

\item{n_players}{Only used if \code{value_func} is a \code{function}.
The number of players in the game.}
}
\value{
The Banzhaf value for each player
}
\description{
Calculate the Banzhaf value
}
