% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banzhaf_owen_appro.R
\name{banzhaf_owen_appro}
\alias{banzhaf_owen_appro}
\title{Banzhaf-Owen Value}
\usage{
banzhaf_owen_appro(characteristic_func, union, n_players, n_rep, replace, echo)
}
\arguments{
\item{characteristic_func}{The valued function defined on the subsets of the number
of players}

\item{union}{List of vectors indicating the a priori unions between the
players}

\item{n_players}{The number of players}

\item{n_rep}{Only used if \code{method} is \code{appro}. The number of
iterations to perform in the approximated calculation.}

\item{replace}{Should sampling be with replacement?}

\item{echo}{Show progress of the calculation.}
}
\value{
The Banzhaf-Owen Index for each player
}
\description{
Calculate the approximated Banzhaf-Owen value using the
algorithm proposed by Saavedra-Nieves & Fiestras-Janeiro (2021).
}
\references{
Saavedra-Nieves, A., & Fiestras-Janeiro, M. G. (2021). Sampling
methods to estimate the Banzhaf–Owen value. Annals of Operations Research,
301(1), 199-223.
}
