% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owen.R
\name{owen}
\alias{owen}
\title{Owen value}
\usage{
owen(
  characteristic_func,
  union,
  method = "exact",
  n_rep = 10000,
  n_players = 0
)
}
\arguments{
\item{characteristic_func}{The valued function defined on the subsets of the number
of players}

\item{union}{List of vectors indicating the apriori unions between the
players}

\item{method}{Method used to calculate the Owen value. Valid methods are:
\code{exact} for the exact value or \code{appro} for approximated polynomial
calculation based on sampling.}

\item{n_rep}{Only used if \code{method} is \code{appro}. The number of
iterations to perform in the approximated calculation}

\item{n_players}{Only used if \code{characteristic_func} is a \code{function}.
The number of players in the game.}
}
\value{
The Owen value for each player
}
\description{
Calculate the Owen value
}
