% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_appro.R
\name{shapley_appro}
\alias{shapley_appro}
\title{Shapley value (approximation)}
\usage{
shapley_appro(characteristic_func, n_players, n_rep, echo)
}
\arguments{
\item{characteristic_func}{The valued function defined on the subsets of the number
of players}

\item{n_players}{The number of players}

\item{n_rep}{The number of iterations to perform in the approximated
calculation}

\item{echo}{Show progress of the calculation.}
}
\value{
The Shapley value for each player
}
\description{
Calculate the approximated Shapley value based on sampling using
the algorithm proposed by Castro et al. (2009).
}
\references{
Castro, J., Gómez, D., & Tejada, J. (2009). Polynomial calculation
of the Shapley value based on sampling. Computers & operations research,
36(5), 1726-1730.
}
